Fixed saving of terrain probabilities between 1 and 0
The MapWriter was saving the floating point value as an int.

+++ src/libtiled/mapwriter.cpp
@@ -313,14 +313,14 @@ void MapWriterPrivate::writeTileset(QXmlStreamWriter &w, const Tileset *tileset,
         const Tile *tile = tileset->tileAt(i);
         const Properties properties = tile->properties();
         unsigned int terrain = tile->terrain();
-        int probability = tile->terrainProbability();
+        float probability = tile->terrainProbability();
 
-        if (!properties.isEmpty() || terrain != 0xFFFFFFFF || probability != -1 || imageSource.isEmpty()) {
+        if (!properties.isEmpty() || terrain != 0xFFFFFFFF || probability != -1.f || imageSource.isEmpty()) {
             w.writeStartElement(QLatin1String("tile"));
             w.writeAttribute(QLatin1String("id"), QString::number(i));
             if (terrain != 0xFFFFFFFF)
                 w.writeAttribute(QLatin1String("terrain"), makeTerrainAttribute(tile));
-            if (probability != -1)
+            if (probability != -1.f)
                 w.writeAttribute(QLatin1String("probability"), QString::number(probability));
             if (!properties.isEmpty())
                 writeProperties(w, properties);
